mitosheet: cleanup file structure and fix bug with stepIDs and creating new graphs
